libnetwork/bitmap: start SetAny from any position
The serial option to the SetAny and SetAnyInRange bitmap operations
is used to find and set the first unset ordinal after the one most
recently set by a SetAny or SetAnyInRange operation instead of finding
the first unset ordinal from the start of the bitmap. There is no way to
customize this policy without extending bitmap. The bitmap needs to
track some hidden state to implement this functionality, which is simply
the ordinal of the last SetAny operation plus one. The SetAny methods
already return the ordinal which was set, so the caller is already
provided with all the necessary information to implement the current
serial SetAny policy!

Drop the hidden state and serial option from bitmap and replace it with
an option for the caller to specify an arbitrary ordinal from which they
want the bitmap to start searching for an unset bit. Give the caller the
flexibility to implement whatever policy they want for updating that
ordinal value. Reimplement the existing serial option inside package
bitseq so existing consumers are unaffected.

Signed-off-by: Cory Snider <csnider@mirantis.com>