libn/d/bridge: generate random MAC addresses
Deriving a stable MAC address from the interface's IP address is a
clever workaround for the cache-invalidation problem with neighbour
containers' ARP caches when the IP address is reassigned. But this
workaround is only viable on single-stack IPv4 container networks.
Pulling the same trick with IPv6 addresses is not viable as a 128-bit
IPv6 address is too long to derive a unique 48-bit MAC address from. The
requirement to keep the IP->MAC mapping stable across reassignments is
even more impractical on a dual-stack container network as the IPv4 and
IPv6 addresses would effectively have to be assigned as a matching pair.
Users have the option to set a static IP or have IPAM allocate one for
each of a container endpoint's IPv4 and IPv6 addresses, so it would be
an unacceptable breaking change to constrain IPv6 address assignment
configuration based on the endpoint's IPv4 address assignment
configuration or vice versa.

There is a standard mechanism to handle cache invalidation when an IP
address is reassigned to a link with a different MAC address: gratuitous
(unsolicted) ARP/ND notifications. Remove the stable MAC address
workaround from the bridge network driver now that unsolicited ARP/ND
notifications are enabled by default on container network interfaces.

Signed-off-by: Cory Snider <csnider@mirantis.com>