libn/osl: stop tracking neighbor entries
The Namespace keeps some state for each inserted neighbor-table entry in
order to delete the entry (and any related entries) given only the IP
and MAC address of the entry to delete. The per-neighbor state required
to delete the entry happens to be the options passed into the
AddNeighbor call and a cached copy of the link name inside the network
namespace, which can be trivially looked up in O(n) time. Drop
neighbor-entry tracking by requiring callers to pass the same arguments
into DeleteNeighbor as they had passed to AddNeighbor.

Signed-off-by: Cory Snider <csnider@mirantis.com>