cmd/docker-proxy: pass open listener to proxy impl
To more cleanly separate the concerns, refactor the TCP and UDP proxies
to be constructed with an existing TCPListener or UDPConn, respectively.
Lift the responsibilty of opening the listener to the entrypoint. Per
the Single Responsibility Principle, this structure affords changing how
the listener is created without having to touch the proxy
implementations. For instance, we could easily make docker-proxy
backwards compatible with older daemon versions (which do not know how
to pass a pre-opened listener fd when launching the proxy) by adding
logic to the entrypoint which opens a listener on the HostAddr:HostPort
supplied as flags if a socket is not passed as fd 4.

Signed-off-by: Cory Snider <csnider@mirantis.com>