daemon/server/middleware: trust APIVersion params
The range of API versions the server accepts is a property of the
daemon. It is outside of the version middleware's scope of
responsibility to assert that the provided range is contained within the
range of versions the daemon (or the api module) supports. Modify the
NewVersionMiddleware constructor to only assert the invariant

    minAPIVersion <= defaultAPIVersion

Signed-off-by: Cory Snider <csnider@mirantis.com>