daemon/server: move APIVersion consts from config
The minimum and maximum API versions supported by the daemon are
static properties of the daemon's implementation: souce code changes are
required to add support for a newer API version or drop support for an
older one. Static properties of a codebase are not runtime configurable
by definition, so it does not make sense to define the API version
constants in the daemon's configuration package. (By contrast, the
minimum API version the daemon will accept requests for is dynamic
configuration whose value is bounded by the minimum-supported and
maximum-supported API versions. And in theory the maximum API version
the daemon will advertise and accept requests for could also be dynamic
configuration which would also be bounded by the same static range.)
Move the MinAPIVersion and MaxAPIVersion constants into the
daemon/server package.

Signed-off-by: Cory Snider <csnider@mirantis.com>